+++ common/bcdmath.cc
@@ -82,7 +82,7 @@ BCD pi()
     return *(const BCDFloat*)(constTable + BCD_CONST_PI);
 }
 
-static void sincosTaylor(const BCD& a, BCD& sa, BCD& ca, int n)
+static void sincosTaylor(const BCD& a, BCD& sa, BCD& ca, int16 n)
 {
     /* calculate sin(a) and cos(a) by taylor series of n terms.
      */
@@ -92,8 +92,8 @@ static void sincosTaylor(const BCD& a, BCD& sa, BCD& ca, int n)
     sa = 1;
     BCD t = 1;
 
-    int i = 1;
-    int j = 1;
+    int16 i = 1;
+    int16 j = 1;
     while (i < n) {
         ++j;
         t *= a2/j;
@@ -117,8 +117,8 @@ void sincos(const BCD& v, BCD* sinv, BCD* cosv)
      */
     BCD res;
     BCD a;
-    int k;
-    int neg = (v < 0);
+    int16 k;
+    int16 neg = (v < 0);
 
     /* arrange a >= 0 */
     if (neg) a = -v;
@@ -159,7 +159,7 @@ void sincos(const BCD& v, BCD* sinv, BCD* cosv)
     BCD sina, cosa;
 
     k &= 64-1;  // wrap around at 2pi
-    int q = k/16; // q is the quadrant.
+    int16 q = k/16; // q is the quadrant.
 
     k &= 15; // index into table.
 
@@ -229,11 +229,11 @@ BCD tan(const BCD& v)
     return s/c;
 }
 
-static BCD expTaylor(const BCD& a, int n)
+static BCD expTaylor(const BCD& a, int16 n)
 {
     BCD t = a;
     BCD s = t + 1;
-    int i = 1;
+    int16 i = 1;
     while (i < n) {
         t = t*a/(++i);
         s += t;
@@ -297,7 +297,7 @@ BCD exp(const BCD& v)
         return 0;
     }
 
-    int k = 64;
+    int16 k = 64;
     BCD r = (v - n*ln2)/k;
     
     /* error bounded by x^9/9! where x = ln(2)/k */
@@ -318,11 +318,11 @@ BCD log(const BCD& v)
     if (v.isZero()) return BCDFloat::negInf();
 
     BCD a = v;
-    int p10 = a.exponent()*4;
+    int16 p10 = a.exponent()*4;
     a.setExponent(0);
 
-    int n;
-    int d = a.digit(0);
+    int16 n;
+    int16 d = a.digit(0);
     if (d >= 1000) { n = 10; p10 -= 1; }
     else if (d >= 100) { n = 100; p10 -= 2; }
     else if (d >= 10) { n = 1000; p10 -= 3; }
@@ -332,7 +332,7 @@ BCD log(const BCD& v)
     a *= n;
 
     BCD oneotwo(*(const BCDFloat*)(constTable + BCD_CONST_ONEOTWO)); // 1.02
-    int p2 = 2;
+    int16 p2 = 2;
 
     while (a >= oneotwo) { 
         p2 <<= 1;
@@ -347,7 +347,7 @@ BCD log(const BCD& v)
     BCD s = t;
     a = t*t;
 
-    int i;
+    int16 i;
     for (i = 3; i < 12; i += 2) { // only 5 terms needed
         t *= a;
         s += t/i;
@@ -389,7 +389,7 @@ BCD atan(const BCD& v)
      *  tan(x/2) = tan(x)/(1+sqrt(1+tan(x)^2))
      */
     BCD atanlim(*(const BCDFloat*)(constTable + BCD_CONST_ATANLIM));
-    int doubles = 0;
+    int16 doubles = 0;
     while (a > atanlim) {
         ++doubles;
         a = a/(1+sqrt(1+a*a));  // at most 3 iterations.
@@ -402,8 +402,8 @@ BCD atan(const BCD& v)
     BCD a2 = a*a;
     BCD t = a2;
     BCD s = 1 - t/3;
-    int i;
-    int j = 5;
+    int16 i;
+    int16 j = 5;
     /* perform 9 more terms, error will be the first term not used.
      * ie x^21/21.
      */
@@ -544,7 +544,7 @@ BCD modtwopi(const BCD& a)
     if (a < pi2) return a;
 
     unsigned short xd[2*P+1];
-    int i;
+    int16 i;
 
     /* copy digits of manstissa as double precision */
     for (i = 0; i < P; ++i) {
@@ -554,11 +554,11 @@ BCD modtwopi(const BCD& a)
         xd[i] = 0;
         ++i;
     }
-    int ex = a.exponent();
+    int16 ex = a.exponent();
 
     unsigned short bd[2*P+1];
-    int eb;
-    int excess = 0;
+    int16 eb;
+    int16 excess = 0;
 
     /* see if the exponent is large enough to consider it separately */
     if (ex > P) {
@@ -572,10 +572,10 @@ BCD modtwopi(const BCD& a)
         ex = P;
 
         unsigned short fd[2*P+1];
-        int ef;
+        int16 ef;
 
         /* do we have enough table? */
-        if (excess + 2*P+1 > (int)(sizeof(inv2pi)/sizeof(inv2pi[0]))) {
+        if (excess + 2*P+1 > (int16)(sizeof(inv2pi)/sizeof(inv2pi[0]))) {
             /* oh dear, digits required off end of table. give up.
              */
             return BCDFloat::nan();
